+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+### Change
+
+- Remove FurucomboProxy event log.
+
 ## [1.0.0] - 2022-05-12
 
 ### Added